requirements.txt
fastapi==0.95.2
graphene==3.2.2
starlette-graphene3==3.0.1
sqlalchemy==1.4.46
argon2-cffi==21.3.0
PyJWT==2.6.0